Add weather package
